fix (admin): don't show rounds on the front page
Refs: TILA-2902
